+++ .ci/coverage.sh
@@ -2,8 +2,8 @@
 set -e -v
 
 if [[ $1 == "script" ]]; then
-  coverage run -m pytest --pyargs nengo
-  coverage run -a -m pytest nengo_dl
+  coverage run -m pytest $TEST_ARGS --pyargs nengo
+  coverage run -a -m pytest $TEST_ARGS nengo_dl
   coverage report -m
 elif [[ $1 == "after_success" ]]; then
   eval "bash <(curl -s https://codecov.io/bash)"