+++ .ci/test.sh
@@ -2,6 +2,6 @@
 set -e -v
 
 if [[ $1 == "script" ]]; then
-  pytest -v -n 2 --color=yes --pyargs nengo
-  pytest -v -n 2 --color=yes --durations 20 $TEST_ARGS nengo_dl
+  pytest $TEST_ARGS -v -n 2 --color=yes --pyargs nengo
+  pytest $TEST_ARGS -v -n 2 --color=yes --durations 20 nengo_dl
 fi